coldata: change Batch interface to operate with uint64 length
This commit changes Batch interface to operate with uint64 length
because it unifies the way we store batches - now we will have a single
implementation MemBatch (with an exception for coldata.ZeroBatch). This
allows us to remove bufferedBatch.

Release note: None

+++ pkg/col/coldata/random_testutils.go
@@ -115,7 +115,7 @@ func RandomVec(
 
 	for i := 0; i < n; i++ {
 		if rng.Float64() < nullProbability {
-			vec.Nulls().SetNull(uint16(i))
+			vec.Nulls().SetNull(uint64(i))
 		}
 	}
 }
